+++ opencompass/models/pangu_api.py
@@ -1,4 +1,3 @@
-import sys
 from concurrent.futures import ThreadPoolExecutor
 from typing import Dict, List, Optional, Union
 
@@ -79,38 +78,6 @@ class PanGu(BaseAPIModel):
                              [max_out_len] * len(inputs)))
         self.flush()
         return results
-
-    def flush(self):
-        """Ensure simultaneous emptying of stdout and stderr when concurrent
-        resources are available.
-
-        When employing multiprocessing with standard I/O redirected to files,
-        it is crucial to clear internal data for examination or prevent log
-        loss in case of system failures."
-        """
-        if hasattr(self, 'tokens'):
-            sys.stdout.flush()
-            sys.stderr.flush()
-
-    def acquire(self):
-        """Acquire concurrent resources if exists.
-
-        This behavior will fall back to wait with query_per_second if there are
-        no concurrent resources.
-        """
-        if hasattr(self, 'tokens'):
-            self.tokens.acquire()
-        else:
-            self.wait()
-
-    def release(self):
-        """Release concurrent resources if acquired.
-
-        This behavior will fall back to do nothing if there are no concurrent
-        resources.
-        """
-        if hasattr(self, 'tokens'):
-            self.tokens.release()
 
     def _get_token(self):
         url = self.token_url
