+++ opencompass/models/sensetime_api.py
@@ -1,4 +1,3 @@
-import sys
 import time
 from concurrent.futures import ThreadPoolExecutor
 from typing import Dict, List, Optional, Union
@@ -18,8 +17,6 @@ class SenseTime(BaseAPIModel):
     Args:
         path (str): The name of SenseTime model.
             e.g. `nova-ptc-xl-v1`
-        model_type (str): The type of the model
-            e.g. `chat`
         key (str): Authorization key.
         query_per_second (int): The maximum queries allowed per second
             between two consecutive calls of the API. Defaults to 1.
@@ -34,8 +31,7 @@ class SenseTime(BaseAPIModel):
         self,
         path: str,
         key: str,
-        model_type: str = 'chat',
-        url: str = 'https://api.sensenova.cn/v1/llm/chat-completions',
+        url: str,
         query_per_second: int = 2,
         max_seq_len: int = 2048,
         meta_template: Optional[Dict] = None,
@@ -50,7 +46,6 @@ class SenseTime(BaseAPIModel):
             'Content-Type': 'application/json',
             'Authorization': f'Bearer {key}'
         }
-        self.type = model_type
         self.url = url
         self.model = path
 
@@ -76,38 +71,6 @@ class SenseTime(BaseAPIModel):
                              [max_out_len] * len(inputs)))
         self.flush()
         return results
-
-    def flush(self):
-        """Ensure simultaneous emptying of stdout and stderr when concurrent
-        resources are available.
-
-        When employing multiprocessing with standard I/O redirected to files,
-        it is crucial to clear internal data for examination or prevent log
-        loss in case of system failures."
-        """
-        if hasattr(self, 'tokens'):
-            sys.stdout.flush()
-            sys.stderr.flush()
-
-    def acquire(self):
-        """Acquire concurrent resources if exists.
-
-        This behavior will fall back to wait with query_per_second if there are
-        no concurrent resources.
-        """
-        if hasattr(self, 'tokens'):
-            self.tokens.acquire()
-        else:
-            self.wait()
-
-    def release(self):
-        """Release concurrent resources if acquired.
-
-        This behavior will fall back to do nothing if there are no concurrent
-        resources.
-        """
-        if hasattr(self, 'tokens'):
-            self.tokens.release()
 
     def _generate(
         self,
